Create Diary of friends
Question:
You need to write classes to maintain a diary of friends.  For each friend in your diary, you are also allowed to write down at most one friend of your friend.  A timestamp is used to keep track of when the friend was created (Hint: use <ctime> and time(NULL) function for timestamp).